projects/sparse-hate-explain/pyproject.toml
[project]
name = "sparse-hate-explain"
version = "0.1.0"
description = "Sparse rationale-constrained attention for faithful hate speech explanations"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.30",
    "datasets>=2.14",
    "captum>=0.6",
    "scikit-learn>=1.3",
    "scipy>=1.10",
    "numpy>=1.24",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "pandas>=2.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "accelerate>=0.20",
    "sentencepiece>=0.1.99",
    "protobuf>=3.20",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff>=0.1"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
